requirements.txt
g4f[all]
discord.py
deep_translator